fix need reader

+++ fluorender/FluoRender/Formats/tif_reader.cpp
@@ -909,7 +909,7 @@ Nrrd* TIFReader::ReadTiff(std::vector<SliceInfo> &filelist,
             m_max_value = max_value;
          else {
             double value;
-            for (int i=0; i<m_slice_num*m_x_size*m_y_size; i++) {
+            for (size_t i=0; i<(size_t)m_slice_num*(size_t)m_x_size*(size_t)m_y_size; i++) {
                value= ((unsigned short*)nrrdout->data)[i];
                m_max_value = value>m_max_value ? value : m_max_value;
             }
